refactor(quiz): split oversized lib/queries/quiz.ts (#852)
quiz.ts was 263 lines — over the 200-line utility cap. Split at the
natural seam into three co-located files (no barrel), each within the
cap:

- quiz-query-types.ts (29L): the shared result types SubjectOption /
  TopicOption / SubtopicOption / TopicWithSubtopics.
- quiz-subject-queries.ts (200L): the internal row types +
  fetchActiveQuestionCounts helper + the 4 subject/topic/subtopic
  lookup functions.
- quiz-session-queries.ts (41L): QuestionFilter + getRandomQuestionIds
  (fully independent of the subject side).

quiz.test.ts split into quiz-subject-queries.test.ts +
quiz-session-queries.test.ts (37 cases preserved, none lost). All 13
importers repointed to the right new file; type-only consumers import
from quiz-query-types, function consumers from the query files;
lookup.ts uses both. lookup.test.ts / start.test.ts vi.mock paths
updated too.

Pure behaviour-preserving move — function bodies byte-identical.
check-types + biome clean; quiz/lib-queries unit suite 1853 green.

Closes #852
