+++ apps/web/app/app/quiz/_hooks/topic-tree-helpers.ts
@@ -1,4 +1,4 @@
-import type { TopicWithSubtopics } from '@/lib/queries/quiz'
+import type { TopicWithSubtopics } from '@/lib/queries/quiz-query-types'
 
 export type UseTopicTreeReturn = {
   topics: TopicWithSubtopics[]
