+++ apps/web/lib/queries/quiz-subject-queries.ts
@@ -1,36 +1,11 @@
 import { createServerSupabaseClient } from '@repo/db/server'
-import type { CalcMode } from '@/app/app/quiz/types'
 import { rpc } from '@/lib/supabase-rpc'
-
-export type SubjectOption = {
-  id: string
-  code: string
-  name: string
-  short: string
-  questionCount: number
-}
-
-export type TopicOption = {
-  id: string
-  code: string
-  name: string
-  questionCount: number
-}
-
-export type SubtopicOption = {
-  id: string
-  code: string
-  name: string
-  questionCount: number
-}
-
-export type TopicWithSubtopics = {
-  id: string
-  code: string
-  name: string
-  questionCount: number
-  subtopics: SubtopicOption[]
-}
+import type {
+  SubjectOption,
+  SubtopicOption,
+  TopicOption,
+  TopicWithSubtopics,
+} from './quiz-query-types'
 
 type SubjectRow = { id: string; code: string; name: string; short: string; sort_order: number }
 type TopicRow = { id: string; code: string; name: string; sort_order: number }
@@ -42,8 +17,6 @@ type QuestionCountRow = {
   // bigint COUNT(*) — PostgREST may serialize it as a string; coerce with Number() at every read site.
   n: number | string
 }
-
-export type QuestionFilter = 'all' | 'unseen' | 'incorrect' | 'flagged'
 
 async function fetchActiveQuestionCounts(
   supabase: Awaited<ReturnType<typeof createServerSupabaseClient>>,
@@ -224,40 +197,4 @@ export async function getTopicsWithSubtopics(subjectId: string): Promise<TopicWi
       subtopics: subtopicsByTopic.get(t.id) ?? [],
     }))
     .filter((t) => t.questionCount > 0)
-}
-
-export async function getRandomQuestionIds(opts: {
-  subjectId: string
-  topicIds?: string[]
-  subtopicIds?: string[]
-  count: number
-  filters?: QuestionFilter[]
-  calcMode?: CalcMode
-}): Promise<string[]> {
-  const supabase = await createServerSupabaseClient()
-  const { subjectId, topicIds, subtopicIds, count, filters, calcMode } = opts
-  const activeFilters = filters?.filter((f) => f !== 'all') ?? []
-
-  // undefined → null to RPC = unconstrained (whole subject pool); [] → empty array = match nothing (topic_id = ANY('{}') is always false).
-  // p_calc_mode is a literal enum the RPC reads directly ('all' = unrestricted via CASE ELSE) — do NOT strip 'all' the way p_filters does.
-  const { data, error } = await rpc<{ id: string }[]>(supabase, 'get_random_question_ids', {
-    p_subject_id: subjectId,
-    p_topic_ids: topicIds ?? null,
-    p_subtopic_ids: subtopicIds ?? null,
-    p_count: count,
-    p_filters: activeFilters,
-    p_calc_mode: calcMode ?? 'all',
-  })
-  if (error) {
-    console.error('[getRandomQuestionIds] get_random_question_ids error:', error.message)
-    return []
-  }
-  if (!Array.isArray(data)) return []
-  // Per-row guard required by code-style.md §5 — the `rpc<{id: string}[]>` cast is
-  // a TypeScript assertion only, not a runtime guarantee. Drop rows that don't
-  // carry a string id; otherwise `undefined` would leak into start_quiz_session's
-  // uuid[] arg and trigger a Postgres type error.
-  return data
-    .map((r) => (r && typeof r === 'object' ? (r as { id?: unknown }).id : undefined))
-    .filter((id): id is string => typeof id === 'string')
 }
